Add HA support for ovnkube masters and for OVN DBs.
This patch adds HA support for ovnkube masters and configures
the OVN ovsdb-servers in active/standby mode so that the OVN db
contents are replicated.

With this patch, we can run multiple ovnkube in master mode.
Only one will be active. The client-go/tools/leaderelection module
is used to elect a leader.

A couple of configuration options are added: Important ones are :-
 - manage-db-servers (boolean) - If set, manages the OVN ovsdb-servers in active/standby
                                 mode. Uses the unix sockets to communicate with
                                 the ovsdb-servers.

 - pod-ip - The IP of pod allocated by k8s. Required when manage-db-servers is enabled.

If manage-db-servers option is set, it is expected that
 - OVN ovsdb-servers and ovn-northd are running in each node where ovnkube-master
   is running, there by tightly coupling all these services in the master node.
 - ovnkube-master can communicate with the ovsdb-servers and ovn-northd using
   the unix control socket. Eg. /var/run/openvswitch/ovnnb_db.ctl
 - ovn-northd is configured to connect to the local ovsdb-servers using the unix
   sockets (and not using the endpoint - ovnkube-db).

If manage-db-servers option is not set,
 - it is expected that the HA of OVN ovsdb-servers if desired is managed separately.
 - OVN ovsdb-servers and ovn-northd can be decoupled from ovnkube-master.

When the ovnkube-master is started, it participates in the
election. The ovnkube-master which wins the election, does the following
   - If manage-db-servers option is set, then
      * Creates or updates the endpoint - ovnkube-db with its IP address.
      * Configures the OVN ovsdb-servers as active, so that the OVN ovsdb-servers
        running in that node allow write transactions.
      * Resumes the ovn-northd running in that node, so that it can process the
        OVN NB DB entries and convert to the logical flows.
   - Starts monitoring the pod events.

There is no change required in ovnkube's running on the nodes as it monitors for
any changes to the ovnkube-db endpoint and updates the ovn-remote option so that
ovn-controller can connect to the active ovsdb-server.

A follow up patch will add the documentation.

Signed-off-by: Numan Siddique <numans@ovn.org>