Remove inefficient inequality operator in firewall match
Egress firewall was using the inequality operator as to not
block internal cluster traffic. However the inequality
operator is a very dangerous operator used in OVN as it
can easily lead to open flow explosions. This is because
ovn-controller needs to create equality defined flows
from that inequality condition, which depending on the cluster
can lead to problems.

We can, in our case, apply some default allow policies for
intra-cluster traffic with the equality operator. Small tests on
KIND clusters have already shown a flow count improvement
using default allow policies.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

Add default allow ACLs for egress firewall

This commit adds default allow ACL rules for egress firewall.
The goal of this commit is to allow ALL in-cluster traffic, that
being:

- Pod -> Pod traffic
- Pod -> Service traffic
- Pod -> host networked pod traffic
- Pod -> Service backed by host networked pod traffic

The last two cases were previously not the case for egress firewall
and hence led to failed connections to the kubernetes default service
for example.

The second is needed because, since we're applying the ACLs on
the logical switches the load balancer DNAT is applied after the
evaluation of the ACLs on the logical switch, hence pod -> service traffic
will be un-NAT-ed when the packet is evaluated for the ACLs

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>