Egress firewall: switch to direction=to-lport for ACLs
The ACLs for egress firewall are currently created while specifying
direction=from-lport, this inhibits egress firewall matching pods in
local GW mode from connecting to services. Specifying to-lport will
allow DNAT to happen and connectivity to services without specifying a
specific ACL allowing traffic to the service CIDR. For shared GW mode
this is indifferent since the ACLs are created on the join switch.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>